Refactor examples
With the introduction of the new Spring Boot starter, the existing
Spring examples get obsolete. A better module names make the code much
more organized.
