+++ engine/src/main/java/de/digitalcollections/flusswerk/engine/exceptions/StopProcessingException.java
@@ -6,9 +6,7 @@ import java.util.function.Supplier;
 /** Throw this exception in case of errors that won't go away when trying again later. */
 public class StopProcessingException extends RuntimeException {
 
-  /**
-   * @param message The message, possibly including format strings
-   */
+  /** @param message The message, possibly including format strings */
   public StopProcessingException(String message) {
     super(message); // String-only constructor is needed for supplier
   }
@@ -43,8 +41,7 @@ public class StopProcessingException extends RuntimeException {
    * @param args Arguments for the format string in <code>message</code>
    * @return A supplier for this exception than can be further customized
    */
-  public static ExceptionSupplier<StopProcessingException> because(
-      String message, Object... args) {
+  public static ExceptionSupplier<StopProcessingException> because(String message, Object... args) {
     return new ExceptionSupplier<>(StopProcessingException.class, message, args);
   }
 }
