+++ examples-plain/src/main/resources/application.yml
