+++ examples-plain/src/main/resources/logback.xml
