+++ examples-spring-boot/src/main/java/de/digitalcollections/flusswerk/spring/boot/example/ComposePerfectGreeting.java
