+++ examples-spring-boot/src/main/resources/application.yml
