--- examples/src/main/java/de/digitalcollections/flusswerk/examples/spring/UppercaseTransformer.java
@@ -1,16 +1,0 @@
-package de.digitalcollections.flusswerk.examples.spring;
-
-import java.util.function.Function;
-import org.springframework.stereotype.Component;
-
-@Component
-public class UppercaseTransformer implements Function<String, String> {
-
-  @Override
-  public String apply(String s) {
-    if (s == null) {
-      return null;
-    }
-    return s.toUpperCase();
-  }
-}
