+++ pom.xml
@@ -53,10 +53,10 @@
 
   <modules>
     <module>engine</module>
-    <module>examples</module>
-    <module>integration</module>
+    <module>examples-plain</module>
+    <module>examples-spring-boot</module>
+    <module>integration-tests</module>
     <module>spring-boot-starter-flusswerk</module>
-    <module>spring-boot-example</module>
   </modules>
 
   <properties>
@@ -71,6 +71,7 @@
     <version.junit>5.6.0</version.junit>
     <version.logback>1.2.3</version.logback>
     <version.logstash>6.3</version.logstash>
+    <version.micrometer>1.3.5</version.micrometer>
     <version.mockito>3.3.0</version.mockito>
     <version.slf4j>1.7.30</version.slf4j>
     <version.spring-boot>2.2.5.RELEASE</version.spring-boot>
